Fixed crash in udInstancedModel_Build() when more than 16 roots are required on a branch layer. Also added udHeapValdate().
This fix is not ideal in that if there are more than 16 roots required on the bottom layer the function will fail early. But if we run out of roots during the build process up the tree it will just discard the leaf nodes that won't fit in that layer.